Added Alpine Linux virt image
This image contains a repacked alpine-virt image. The original file is
alpine-virt-3.8.1-x86_64.iso . I have simply copied all the content
into an ext2 filesystem. This approach has gaps, but it is a start to
add integration tests. In the future this will be improved, and
potentially integrated with GitHub's deployments on our own infra.